convert float16b to float16

+++ G-retriever/models/graph_llm.py
@@ -108,7 +108,7 @@ class GraphLLM(torch.nn.Module):
     def device(self):
         return list(self.parameters())[0].device
 
-    def maybe_autocast(self, dtype=torch.bfloat16):#reduces memory usage and accelerates computations on GPUs.
+    def maybe_autocast(self, dtype=torch.float16):#reduces memory usage and accelerates computations on GPUs.
         # if on cpu, don't use autocast
         # if on gpu, use autocast with dtype if provided, otherwise use torch.float16
         enable_autocast = self.device != torch.device("cpu")
